requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython>=2.0.0
email-validator==2.1.0.post1
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
motor==3.3.2
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
pymongo==4.6.2
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
